Completes user story https://git.cardiff.ac.uk/c21052928/energy-data-centre-visualisation/-/issues/6
I have made a component for the bar-chart that is reusable, requiring the input of a few options/props. This component is used here to create a bar-chart, visualising the data regarding the Thermal_characteristics related CSVs.

Showing the researcher the annual heat demand before and after energy improvements have been made per heating technology.

I migrated from d3.js to chart.js as it proves to be signficantly easier to use, and looks visually pleasing.